+++ src/ext_http.c
@@ -1038,7 +1038,7 @@ static void handle_request(int fd) {
                     TokenList auth_tl = tokenize(auth_src);
                     ASTNode *auth_ast = parse(&auth_tl);
                     Env *auth_env = env_new(g_global_env);
-                    EigsChunk *auth_chunk = compile_ast(auth_ast, auth_env);
+                    EigsChunk *auth_chunk = compile_ast(auth_ast, auth_env, auth_src);
                     Value *auth_result = vm_execute(auth_chunk, auth_env);
                     chunk_free(auth_chunk);
                     env_decref(auth_env);
@@ -1058,7 +1058,7 @@ static void handle_request(int fd) {
                 TokenList tl = tokenize(r->payload);
                 ASTNode *ast = parse(&tl);
                 Env *req_env = env_new(g_global_env);
-                EigsChunk *req_chunk = compile_ast(ast, req_env);
+                EigsChunk *req_chunk = compile_ast(ast, req_env, r->payload);
                 Value *result = vm_execute(req_chunk, req_env);
                 chunk_free(req_chunk);
                 char *result_str = value_to_string(result);